feat: created lib for records and enums in common with all services
