--- inventory-service/src/main/java/br/com/microservices/orchestrated/inventoryservice/core/dto/Event.java
@@ -1,18 +1,0 @@
-package br.com.microservices.orchestrated.inventoryservice.core.dto;
-
-import br.com.microservices.orchestrated.inventoryservice.core.enums.SagaStatus;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public record Event(
-        String id,
-        String transactionId,
-        String orderId,
-        Order payload,
-        String source,
-        SagaStatus status,
-        List<History> eventHistory,
-        LocalDateTime createdAt
-) {
-}
