--- inventory-service/src/main/java/br/com/microservices/orchestrated/inventoryservice/core/dto/OrderProduct.java
@@ -1,9 +1,0 @@
-package br.com.microservices.orchestrated.inventoryservice.core.dto;
-
-import br.com.microservices.orchestrated.orchestratorservice.core.dto.Product;
-
-public record OrderProduct(
-        Product product,
-        Integer quantity
-) {
-}
