--- inventory-service/src/main/java/br/com/microservices/orchestrated/inventoryservice/core/dto/Product.java
@@ -1,7 +1,0 @@
-package br.com.microservices.orchestrated.inventoryservice.core.dto;
-
-public record Product(
-        String code,
-        Double unitValue
-) {
-}
