--- inventory-service/src/main/java/br/com/microservices/orchestrated/inventoryservice/core/enums/SagaStatus.java
@@ -1,4 +1,0 @@
-package br.com.microservices.orchestrated.inventoryservice.core.enums;
-
-public enum SagaStatus {
-}
