--- orchestrator-service/src/main/java/br/com/microservices/orchestrated/orchestratorservice/core/dto/OrderProduct.java
@@ -1,7 +1,0 @@
-package br.com.microservices.orchestrated.orchestratorservice.core.dto;
-
-public record OrderProduct(
-        Product product,
-        Integer quantity
-) {
-}
