--- orchestrator-service/src/main/java/br/com/microservices/orchestrated/orchestratorservice/core/enums/EventSource.java
@@ -1,8 +1,0 @@
-package br.com.microservices.orchestrated.orchestratorservice.core.enums;
-
-public enum EventSource {
-    ORCHESTRATOR,
-    PRODUCT_VALIDATION_SERVICE,
-    PAYMENT_SERVICE,
-    INVENTORY_SERVICE
-}
