--- orchestrator-service/src/main/java/br/com/microservices/orchestrated/orchestratorservice/core/enums/Topic.java
@@ -1,25 +1,0 @@
-package br.com.microservices.orchestrated.orchestratorservice.core.enums;
-
-public enum Topic {
-    START_SAGA("start-saga"),
-    BASE_ORCHESTRATOR("orchestrator"),
-    FINISH_SUCCESS("finish-success"),
-    FINISH_FAIL("finish-fail"),
-    PRODUCT_VALIDATION_SUCCESS("product-validation-success"),
-    PRODUCT_VALIDATION_FAIL("product-validation-fail"),
-    PAYMENT_SUCCESS("payment-success"),
-    PAYMENT_FAIL("payment-fail"),
-    INVENTORY_SUCCESS("inventory-success"),
-    INVENTORY_FAIL("inventory-fail"),
-    NOTIFY_ENDING("notify-ending");
-
-    private final String topic;
-
-    Topic(String topic) {
-        this.topic = topic;
-    }
-
-    public String getTopic() {
-        return topic;
-    }
-}
