--- payment-service/src/main/java/br/com/microservices/orchestrated/paymentservice/core/dto/Product.java
@@ -1,7 +1,0 @@
-package br.com.microservices.orchestrated.paymentservice.core.dto;
-
-public record Product(
-        String code,
-        Double unitValue
-) {
-}
