--- payment-service/src/main/java/br/com/microservices/orchestrated/paymentservice/core/enums/SagaStatus.java
@@ -1,4 +1,0 @@
-package br.com.microservices.orchestrated.paymentservice.core.enums;
-
-public enum SagaStatus {
-}
