--- product-validation-service/src/main/java/br/com/microservices/orchestrated/productvalidationservice/core/enums/SagaStatus.java
@@ -1,7 +1,0 @@
-package br.com.microservices.orchestrated.productvalidationservice.core.enums;
-
-public enum SagaStatus {
-    SUCCESS,
-    ROLLBACK_PENDING,
-    FAIL
-}
