+++ saga-utils/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Mar 10 11:48:33 BRT 2025
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-8.10-bin.zip
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
