+++ saga-utils/src/main/java/io/github/paulovieirajr/dto/Order.java
@@ -1,0 +1,14 @@
+package io.github.paulovieirajr.dto;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public record Order(
+        String id,
+        List<OrderProduct> products,
+        LocalDateTime createdAt,
+        String transactionId,
+        Double totalAmount,
+        Integer totalItems
+) {
+}
