+++ saga-utils/src/main/java/io/github/paulovieirajr/dto/OrderProduct.java
@@ -1,0 +1,7 @@
+package io.github.paulovieirajr.dto;
+
+public record OrderProduct(
+        Product product,
+        Integer quantity
+) {
+}
