+++ saga-utils/src/main/java/io/github/paulovieirajr/dto/Product.java
@@ -1,0 +1,7 @@
+package io.github.paulovieirajr.dto;
+
+public record Product(
+        String code,
+        Double unitValue
+) {
+}
