+++ saga-utils/src/main/java/io/github/paulovieirajr/dto/services/Event.java
@@ -1,0 +1,18 @@
+package io.github.paulovieirajr.dto.services;
+
+import io.github.paulovieirajr.dto.Order;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public record Event(
+        String id,
+        String transactionId,
+        String orderId,
+        Order payload,
+        String source,
+        String status,
+        List<History> eventHistory,
+        LocalDateTime createdAt
+) {
+}
