+++ saga-utils/src/main/java/io/github/paulovieirajr/dto/services/History.java
@@ -1,0 +1,11 @@
+package io.github.paulovieirajr.dto.services;
+
+import java.time.LocalDateTime;
+
+public record History(
+        String source,
+        String status,
+        String message,
+        LocalDateTime createdAt
+) {
+}
