+++ saga-utils/src/main/java/io/github/paulovieirajr/enums/EventSource.java
@@ -1,0 +1,8 @@
+package io.github.paulovieirajr.enums;
+
+public enum EventSource {
+    ORCHESTRATOR,
+    PRODUCT_VALIDATION_SERVICE,
+    PAYMENT_SERVICE,
+    INVENTORY_SERVICE
+}
