+++ saga-utils/src/main/java/io/github/paulovieirajr/enums/SagaStatus.java
@@ -1,0 +1,4 @@
+package io.github.paulovieirajr.enums;
+
+public enum SagaStatus {
+}
