+++ saga-utils/src/main/java/io/github/paulovieirajr/enums/Topic.java
@@ -1,0 +1,25 @@
+package io.github.paulovieirajr.enums;
+
+public enum Topic {
+    START_SAGA("start-saga"),
+    BASE_ORCHESTRATOR("orchestrator"),
+    FINISH_SUCCESS("finish-success"),
+    FINISH_FAIL("finish-fail"),
+    PRODUCT_VALIDATION_SUCCESS("product-validation-success"),
+    PRODUCT_VALIDATION_FAIL("product-validation-fail"),
+    PAYMENT_SUCCESS("payment-success"),
+    PAYMENT_FAIL("payment-fail"),
+    INVENTORY_SUCCESS("inventory-success"),
+    INVENTORY_FAIL("inventory-fail"),
+    NOTIFY_ENDING("notify-ending");
+
+    private final String topic;
+
+    Topic(String topic) {
+        this.topic = topic;
+    }
+
+    public String getTopic() {
+        return topic;
+    }
+}
