+++ saga-utils/src/main/java/io/github/paulovieirajr/exception/ValidationException.java
@@ -1,0 +1,7 @@
+package io.github.paulovieirajr.exception;
+
+public class ValidationException extends RuntimeException {
+    public ValidationException(String message) {
+        super(message);
+    }
+}
